% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_default_labels.R
\name{add_default_labels}
\alias{add_default_labels}
\title{Add default variable labels}
\usage{
add_default_labels(data)
}
\arguments{
\item{data}{a \link{data.frame}}
}
\value{
a modified version of \code{data}
}
\description{
Add default variable labels
}
\examples{
iris |>
  add_default_labels() |>
  labelled::var_label(unlist = TRUE)
}
