% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels.R
\name{add_labels}
\alias{add_labels}
\title{Add labels to biomarkers}
\usage{
add_labels(data)
}
\arguments{
\item{data}{a \link{data.frame} containing the relevant columns}
}
\value{
a modified version of \code{data} with var labels added
}
\description{
Add labels to biomarkers
}
\examples{
test_data |> add_labels()
}
