% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_range_medians.R
\name{age_range_medians}
\alias{age_range_medians}
\title{replace decade codes with medians of those decades}
\usage{
age_range_medians(x)
}
\arguments{
\item{x}{a \link{character} vector}
}
\value{
a \link{numeric} vector corresponding to \code{x}
}
\description{
replace decade codes with medians of those decades
}
