% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_potential_biomarkers.R
\name{all_potential_biomarkers}
\alias{all_potential_biomarkers}
\title{Get a list of all potential biomarkers available in dataset}
\usage{
all_potential_biomarkers(dataset)
}
\arguments{
\item{dataset}{\link{data.frame} Dataset to extract names from}
}
\value{
a \link{list}
}
\description{
Get a list of all potential biomarkers available in dataset
}
