% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{autoplot.permutation_test}
\alias{autoplot.permutation_test}
\title{Plot the distribution of a permuted test statistic}
\usage{
\method{autoplot}{permutation_test}(object, ...)
}
\arguments{
\item{object}{an object of class \verb{"permutation_test}}

\item{..., }{additional arguments (not used)}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} constructing a histogram of the test statistic
}
\description{
Plot the distribution of a permuted test statistic
}
