% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PF.R
\name{autoplot.PF}
\alias{autoplot.PF}
\title{Create positional variance diagram (PVD)}
\usage{
\method{autoplot}{PF}(object, size.y = 10, color_label = "Pr(stage)", ...)
}
\arguments{
\item{object}{a "PF" object (created by \code{compute_position_frequencies()})}

\item{size.y}{size of biomarker event labels}

\item{color_label}{label for legend color scale}

\item{...}{not currently used}
}
\value{
a "PVD" object (extends \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}})
}
\description{
Create positional variance diagram (PVD)
}
