% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_abbrevs_caption.R
\name{build_abbrevs_caption}
\alias{build_abbrevs_caption}
\title{Build abbreviations caption}
\usage{
build_abbrevs_caption(abbreviations, abbreviations_to_define)
}
\arguments{
\item{abbreviations}{a \link{data.frame} loaded from \code{"inst/analyses/abbreviations.csv"}}

\item{abbreviations_to_define}{a \link{character} \link{vector} indicating rows of \code{abbreviations} to use}
}
\value{
a \link{character} string
}
\description{
Build abbreviations caption
}
\examples{
\dontrun{
abbreviations <- read.csv("inst/analyses/abbreviations.csv")

abbrevs_caption <- abbreviations |>
 build_abbrevs_caption(
   abbreviations_to_define = c(
     "SWM",
     "MRI"
   )
 )
}
}
