% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_abstract_plot.R
\name{build_abstract_plot}
\alias{build_abstract_plot}
\title{Build ggplot with abstract text for cowplot}
\usage{
build_abstract_plot(abstract_text, abstract_size = 5, abstract_width = 50)
}
\arguments{
\item{abstract_text}{\link{character}
String that summarizes the study's main findings.}

\item{abstract_size}{\link{numeric} Abstract text size}

\item{abstract_width}{\link{numeric} Length at which to wrap the abstract text.}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Build ggplot with abstract text for cowplot
}
