% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_rds_path.R
\name{build_rds_path}
\alias{build_rds_path}
\title{Find path to RDS file with output}
\usage{
build_rds_path(dataset_name, output_folder)
}
\arguments{
\item{dataset_name}{\code{\link[=character]{character()}}}

\item{output_folder}{\code{\link[=character]{character()}}}
}
\value{
a \code{\link[=character]{character()}} variable
}
\description{
Find path to RDS file with output
}
\examples{
build_rds_path("sample_data", "output/output.fixed_CV")
}
