% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_score_vals.R
\name{build_score_vals}
\alias{build_score_vals}
\title{Construct the table indicating how many levels each biomarker has}
\usage{
build_score_vals(biomarker_levels)
}
\arguments{
\item{biomarker_levels}{a list containing the levels for each biomarker}
}
\value{
a \code{\link[=matrix]{matrix()}} indicating which levels each biomarker has
}
\description{
Construct the table indicating how many levels each biomarker has
}
