% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_gender.R
\name{clean_gender}
\alias{clean_gender}
\title{Clean the variable \code{Gender}}
\usage{
clean_gender(data)
}
\arguments{
\item{data}{a \link{data.frame}}
}
\value{
a modified version of \code{data}
}
\description{
This function attempts to add the label "chromosomal sex",
but lots of functions remove attributes, such as \code{\link[=droplevels]{droplevels()}}
}
\examples{
test_data |> clean_gender()
}
