% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_mol_act_ratio.R
\name{clean_mol_act_ratio}
\alias{clean_mol_act_ratio}
\title{Clean \code{mol_act_ratio} variable from gp4}
\usage{
clean_mol_act_ratio(x)
}
\arguments{
\item{x}{\link{character} the \code{mol_act_ratio} variable from gp4}
}
\value{
a \link{numeric} version of \code{x}
}
\description{
Clean \code{mol_act_ratio} variable from gp4
}
\examples{
mol_act_ratio <- c(NA, "0.19", "N/A", "0.9*")

mol_act_ratio |> clean_mol_act_ratio()
}
