% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_numeric.R
\name{clean_numeric}
\alias{clean_numeric}
\title{Title}
\usage{
clean_numeric(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{
  Arguments passed on to \code{\link[=missing_codes_to_NA]{missing_codes_to_NA}}
  \describe{
    \item{\code{NA_codes}}{numeric values that should be turned into to NAs}
  }}
}
\description{
Title
}
