% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_mri_levels.R
\name{collapse_mri_level}
\alias{collapse_mri_level}
\title{Collapse levels of a single MRI variable}
\usage{
collapse_mri_level(x, levels = c("Moderate", "Severe"))
}
\arguments{
\item{x}{a MRI variable (\link{factor})}

\item{levels}{\link{character} string vector of levels to collapse}
}
\value{
a modified version of \code{x}
}
\description{
Collapse levels of a single MRI variable
}
\keyword{internal}
