% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_mri_levels.R
\name{collapse_mri_levels}
\alias{collapse_mri_levels}
\title{Collapse MRI levels}
\usage{
collapse_mri_levels(dataset, ...)
}
\arguments{
\item{dataset}{a \link[tibble:tbl_df]{tibble::tbl_df}}

\item{...}{
  Arguments passed on to \code{\link{collapse_mri_level}}
  \describe{
    \item{\code{x}}{a MRI variable (\link{factor})}
    \item{\code{levels}}{\link{character} string vector of levels to collapse}
  }}
}
\value{
a modified version of \code{dataset}
}
\description{
Collapse MRI levels
}
