% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_scid_levels.R
\name{collapse_scid_level}
\alias{collapse_scid_level}
\title{Collapse levels of a single scid variable}
\usage{
collapse_scid_level(
  x,
  levels = c("Absent", "Sub-Threshold"),
  new_level = paste(levels, collapse = "/")
)
}
\arguments{
\item{x}{a scid variable (\link{factor})}

\item{levels}{\link{character} string vector of levels to collapse}

\item{newlevel}{\link{character} string for new collapsed level name}
}
\value{
a modified version of \code{x}
}
\description{
Collapse levels of a single scid variable
}
\keyword{internal}
