% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_scid_levels.R
\name{collapse_scid_levels}
\alias{collapse_scid_levels}
\title{Collapse SCID domain levels}
\usage{
collapse_scid_levels(dataset, ...)
}
\arguments{
\item{dataset}{a \link[tibble:tbl_df]{tibble::tbl_df}}

\item{...}{
  Arguments passed on to \code{\link{collapse_scid_level}}
  \describe{
    \item{\code{x}}{a scid variable (\link{factor})}
    \item{\code{levels}}{\link{character} string vector of levels to collapse}
    \item{\code{new_level}}{\link{character} string for new collapsed level name}
  }}
}
\value{
a modified version of \code{dataset}
}
\description{
Collapse SCID domain levels
}
