% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_permutation_test_stats.R
\name{collect_permutation_test_stats}
\alias{collect_permutation_test_stats}
\title{Collect permutation test statistics}
\usage{
collect_permutation_test_stats(
  output_folder = fs::path(here::here(), "output/output.fixed_CV"),
  permutation_results = fs::path(output_folder, "permutation_test_stats")
)
}
\arguments{
\item{permutation_results}{}
}
\description{
Collect permutation test statistics
}
\examples{
\dontrun{
test_stats = collect_permutation_test_stats("output/output.fixed_CV")
}
}
