% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_subtype_and_demos.R
\name{combine_subtype_and_demos}
\alias{combine_subtype_and_demos}
\title{Combine demographic data with latent subtype classifications}
\usage{
combine_subtype_and_demos(patient_data, subtype_data)
}
\arguments{
\item{patient_data}{a \link{data.frame} containing patient demographics}

\item{subtype_data}{a \code{SuSt_table}
(produced as a list element by \code{\link[=extract_results_from_pickle]{extract_results_from_pickle()}})}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df} combining \code{patient_data} and \code{subtype_data}
via \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}}
(assuming they were provided in the same row-order)
}
\description{
was originally going to be more useful, but got reduced to spaghetti.
probably can be deleted
}
\examples{
sim_data |>
  dplyr::filter(Category == "Patient") |>
  combine_subtype_and_demos(
    subtype_data = sim_subtype_and_stage_table
  )
}
\keyword{internal}
