% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tremor.R
\name{combine_tremor}
\alias{combine_tremor}
\title{Combine exam and history tremor vars}
\usage{
combine_tremor(
  dataset,
  tremor_vars = c("intention tremor hx", "resting tremor hx", "postural tremor hx",
    "intention tremor exam", "resting tremor exam", "postural tremor exam",
    "head tremor hx", "head tremor exam")
)
}
\arguments{
\item{dataset}{\link{data.frame}}

\item{tremor_vars}{\link{character} \link{vector}}
}
\value{
a modified version of \code{dataset}
}
\description{
Tremor variables have TWO versions: "ne_tremor" and "neu_tremor"
ne = Neurological Exam, neu = Neurological History
Here, we create a new Tremor variable that takes the more severe score
(e.g. Yes)
}
\examples{
\dontrun{
test_data |> combine_tremor()
}
}
