% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_pvd_data_prep.R
\name{compact_pvd_data_prep}
\alias{compact_pvd_data_prep}
\title{plot_compact_pvd: data prep}
\usage{
compact_pvd_data_prep(
  figs,
  biomarker_order = NULL,
  biomarker_var = "biomarker_label"
)
}
\arguments{
\item{figs}{a \link{list} of \code{PVD} objects}

\item{biomarker_order}{a \link{character} vector specifying the
order of biomarkers to use}

\item{biomarker_var}{either "biomarker" or "biomarker_label"}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
plot_compact_pvd: data prep
}
\keyword{internal}
