% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_pvd_est_data_prep.R
\name{compact_pvd_est_data_prep}
\alias{compact_pvd_est_data_prep}
\title{plot_compact_pvd_est: data prep}
\usage{
compact_pvd_est_data_prep(figs)
}
\arguments{
\item{figs}{a \link{list} of PVDs}
}
\description{
plot_compact_pvd_est: data prep
}
