% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_pvd_facet_labels.R
\name{compact_pvd_facet_labels}
\alias{compact_pvd_facet_labels}
\title{Create facet labels}
\usage{
compact_pvd_facet_labels(figs, facet_label_prefix)
}
\arguments{
\item{figs}{a \link{list} of PVDs}

\item{facet_label_prefix}{a \link{character} string (length 1)}
}
\description{
Create facet labels
}
