% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_biomarker_groups_table.R
\name{compile_biomarker_groups_table}
\alias{compile_biomarker_groups_table}
\title{Compile biomarker groups table}
\usage{
compile_biomarker_groups_table(
  dataset,
  biomarker_group_list = compile_biomarker_group_list(dataset, ...),
  colors = choose_biomarker_group_colors(biomarker_group_list),
  ...
)
}
\arguments{
\item{dataset}{passed to \code{\link[=compile_biomarker_group_list]{compile_biomarker_group_list()}}}

\item{biomarker_group_list}{todo}

\item{colors}{which colors to use}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
Compile biomarker groups table
}
\examples{
biomarker_group_list <- list(
  "group 1" = c("Biomarker 1", "Biomarker 2"),
  "group 2" = c("Biomarker 3", "Biomarker 4"),
  "group 3" = "Biomarker 5"
)

compile_biomarker_groups_table(
  dataset = sim_data,
  biomarker_group_list = biomarker_group_list
)

}
