% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_correct2.R
\name{compute_prob_correct2}
\alias{compute_prob_correct2}
\title{Title}
\usage{
compute_prob_correct2(dataset, biomarkers, DataScores, max_prob = 1)
}
\arguments{
\item{dataset}{Should be just the control data?}

\item{max_prob}{}
}
\description{
Title
}
