% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_dist2.R
\name{compute_prob_dist2}
\alias{compute_prob_dist2}
\title{Title}
\usage{
compute_prob_dist2(ModelScores, DataScores, biomarkers, prob_correct)
}
\arguments{
\item{prob_correct}{}
}
\description{
Title
}
