% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_scores.R
\name{compute_prob_scores}
\alias{compute_prob_scores}
\title{Title}
\usage{
compute_prob_scores(
  dataset,
  biomarkers,
  ModelScores,
  DataScores,
  prob_dist,
  patient_data = select(filter(dataset, `FX*` == "CGG >= 55"), all_of(biomarkers))
)
}
\arguments{
\item{prob_dist}{}
}
\description{
Title
}
