% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_score_levels.R
\name{compute_score_levels}
\alias{compute_score_levels}
\title{Title}
\usage{
compute_score_levels(dataset)
}
\arguments{
\item{dataset}{a \link{data.frame}}
}
\value{
a \link{character} \link{vector} of levels
}
\description{
not sure if we need this anymore
}
