% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_biomarker_events_table.R
\name{construct_biomarker_events_table}
\alias{construct_biomarker_events_table}
\title{Title}
\usage{
construct_biomarker_events_table(biomarker_levels, biomarker_groups)
}
\arguments{
\item{biomarker_levels}{todo}

\item{biomarker_groups}{todo}
}
\value{
a \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Title
}
