% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts_and_pcts.R
\name{counts_and_pcts}
\alias{counts_and_pcts}
\title{Compute and paste counts and proportions}
\usage{
counts_and_pcts(x)
}
\arguments{
\item{x}{a \link{vector}}
}
\value{
a \link{character} string
}
\description{
Compute and paste counts and proportions
}
\examples{
x <- test_data$`Ataxia: severity*`
x |> counts_and_pcts()

}
