% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scid_domains.R
\name{create_scid_domains}
\alias{create_scid_domains}
\title{Create SCID domain variables}
\usage{
create_scid_domains(dataset, scid_vars = all_scid_vars)
}
\arguments{
\item{dataset}{a \link[tibble:tbl_df]{tibble::tbl_df} containing all of \code{scid_vars} as columns}

\item{scid_vars}{a \code{\link[=character]{character()}} vector}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
Create SCID domain variables
}
