% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cases_and_controls.R
\name{define_cases_and_controls}
\alias{define_cases_and_controls}
\title{Categorize `CGG (backfilled)``}
\usage{
define_cases_and_controls(dataset)
}
\arguments{
\item{dataset}{a tibble containing a numeric variable named \code{CGG (backfilled)}}
}
\value{
a tibble with the same columns as \code{dataset}, except with the following additions:
\itemize{
\item FX (logical): TRUE if CGG >= 55
}
\itemize{
\item \verb{FX*} (factor): dichotomizes CGG to < 55 vs >= 55; \code{NA}s stay as \code{NA}
\item \verb{FX**} (factor): dichotomizes CGG to < 55 vs >= 55; \code{NA}s converted to level "CGG missing"
\item \verb{FX***} (factor): trichotomizes CGG to < 55, 55-100, and >100 >= 55; \code{NA}s converted to level "CGG missing"
}
}
\description{
Adds categorized versions of CGG to input dataset.
}
