% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_order_heatmap.R
\name{event_order_heatmap}
\alias{event_order_heatmap}
\title{Event order heatmap}
\usage{
event_order_heatmap(samples, biomarkers = attr(samples, "biomarker labels"))
}
\arguments{
\item{samples}{a list of orderings}

\item{biomarkers)}{}
}
\value{
a ggplot object that constructs a heatmap
}
\description{
Event order heatmap
}
