% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_likelihoods.R
\docType{data}
\name{example_likelihoods}
\alias{example_likelihoods}
\title{example likelihoods matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 100000 rows and 3 columns.
}
\source{
\code{fxtas/data-raw/example_likelihoods.R}
}
\usage{
example_likelihoods
}
\description{
an example likelihoods matrix
}
\keyword{datasets}
