% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cvic.R
\name{extract_cvic}
\alias{extract_cvic}
\title{Extract CVIC element from an object}
\usage{
extract_cvic(object, ...)
}
\arguments{
\item{object}{a \code{"SuStaIn_output"} object (from \code{\link[=run_OSA]{run_OSA()}})}

\item{...}{unused}
}
\value{
a "CVIC" object
}
\description{
Extract CVIC element from an object
}
\examples{
osa_output_sim_data |> extract_cvic()
}
