% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_permuted_likelihoods.R
\name{extract_permuted_likelihoods}
\alias{extract_permuted_likelihoods}
\title{Extract permuted log-likelihoods from pickle files}
\usage{
extract_permuted_likelihoods(
  permuting_variables = "Gender",
  n_permutations = 1000,
  permutations = 1:n_permutations,
  output_folder,
  verbose = TRUE
)
}
\arguments{
\item{permuting_variables}{which files were permuted}

\item{n_permutations}{how many permutations were performed}

\item{permutations}{which permutation seeds to extract}

\item{output_folder}{where to find the pickle files}

\item{verbose}{whether to print informative messages about progress}
}
\value{
a named \code{\link[=numeric]{numeric()}} vector of log-likelihoods from permuted data
}
\description{
Extract permuted log-likelihoods from pickle files
}
