% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results_from_pickle.R
\name{extract_results_from_pickles}
\alias{extract_results_from_pickles}
\title{Extract results from multiple pickle files}
\usage{
extract_results_from_pickles(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  rda_filename = "data.RData",
  picklename = paste0(dataset_name, "_subtype", n_s - 1, ".pickle"),
  ...
)
}
\arguments{
\item{n_s}{helps construct \code{picklename} by identifying number of latent subgroups}

\item{dataset_name}{root name of dataset}

\item{output_folder}{where to find the dataset}

\item{rda_filename}{name of rda file containing environment used to run analyses}

\item{picklename}{the name of the pickle file to open}
}
\description{
Extract results from multiple pickle files
}
