% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_biomarkers_table.R
\name{flex_biomarkers_table}
\alias{flex_biomarkers_table}
\title{Print method for \code{biomarkers_table}}
\usage{
flex_biomarkers_table(x, width_biomarkers = 1.5)
}
\arguments{
\item{x}{a "biomarkers_table" object from \code{make_biomarkers_table()}}

\item{width_biomarkers}{\link{numeric}: width for \code{Symptom} column}
}
\value{
a \code{\link[flextable:flextable]{flextable::flextable()}}
}
\description{
Print method for \code{biomarkers_table}
}
