% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_demographics_table_as_flextable.R
\name{format_demographics_table_as_flextable}
\alias{format_demographics_table_as_flextable}
\title{Format the demographics table as a flextable}
\usage{
format_demographics_table_as_flextable(x, ...)
}
\arguments{
\item{x}{a \link[gtsummary:gtsummary]{gtsummary::gtsummary} object}

\item{...}{
  Arguments passed on to \code{\link{shared_flextable_settings}}
  \describe{
    \item{\code{padding}}{controls \code{padding.top} and \code{padding.bottom} in \code{\link[flextable:padding]{flextable::padding()}}}
  }}
}
\value{
a \link[flextable:flextable]{flextable::flextable}
}
\description{
Format the demographics table as a flextable
}
\keyword{internal}
