% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_likelihoods.R
\name{format_likelihoods}
\alias{format_likelihoods}
\title{Title}
\usage{
format_likelihoods(likelihoods)
}
\arguments{
\item{likelihoods}{todo}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
Title
}
