% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_results_list.R
\name{format_results_list}
\alias{format_results_list}
\title{Format results list extracted from pickle file}
\usage{
format_results_list(
  results,
  biomarker_labels = names(biomarker_levels),
  biomarker_levels = NULL,
  biomarker_events_table = get_biomarker_events_table(biomarker_levels),
  biomarker_event_names = biomarker_events_table$biomarker_level,
  format_sst = TRUE,
  n_s = get_n_subtypes(results),
  subtype_order = seq_len(n_s),
  biomarker_groupss
)
}
\arguments{
\item{results}{pickle file contents}

\item{biomarker_labels}{biomarker labels}

\item{biomarker_levels}{biomarker levels}

\item{biomarker_events_table}{table of biomarker events
(excluding base level)}

\item{biomarker_event_names}{vector of biomarker event names}

\item{format_sst}{should the subtype and stage table be formatted?
(doesn't work for cross-validation fold pickle-files)}

\item{n_s}{number of latent subgroups}

\item{subtype_order}{\link{integer} indicating how to reorder the subtypes}

\item{biomarker_groups}{a \link{list}}
}
\value{
a \code{"SuStaIn_model"} object (extends \code{\link[=list]{list()}})
}
\description{
Format results list extracted from pickle file
}
