% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxtas-package.R
\docType{package}
\name{fxtas-package}
\alias{fxtas}
\alias{fxtas-package}
\title{fxtas: FXTAS}
\description{
PKG_DESC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ucd-iddrc/fxtas}
  \item \url{https://ucd-iddrc.github.io/fxtas/}
  \item Report bugs at \url{https://github.com/ucd-iddrc/fxtas/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu}

Authors:
\itemize{
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu}
  \item Matthew Ponzini \email{mdponzini@ucdavis.edu}
}

}
\keyword{internal}
