% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_event_names.R
\name{get_biomarker_event_names}
\alias{get_biomarker_event_names}
\title{Extract biomarker event names}
\usage{
get_biomarker_event_names(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{additional arguments passed to class-specific methods}
}
\value{
a \link{character} \link{vector}
}
\description{
Extract biomarker event names
}
