% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_events_table.R
\name{get_biomarker_events_table}
\alias{get_biomarker_events_table}
\title{Construct biomarker events table}
\usage{
get_biomarker_events_table(biomarker_levels, do_factor = TRUE)
}
\arguments{
\item{biomarker_levels}{todo}

\item{do_factor}{whether to turn \code{biomarker} column into factor and sort on it, after \code{level}}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
Construct biomarker events table
}
\examples{
df <- test_data_v1
biomarker_groups <- compile_biomarker_groups_table(dataset = df)
biomarker_varnames <-
  biomarker_groups |>
  dplyr::pull("biomarker")

biomarker_levels <-
  df |>
  get_levels(biomarker_varnames, keep_labels = TRUE)

get_biomarker_events_table(biomarker_levels)
}
