% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_levels.R
\name{get_levels}
\alias{get_levels}
\title{Extract levels from a set of variable names}
\usage{
get_levels(data, varnames)
}
\arguments{
\item{data}{a \code{\link[=data.frame]{data.frame()}}}

\item{varnames}{a \code{\link[=character]{character()}} specifying which columns of \code{data} to analyze}
}
\value{
a \code{\link[=list]{list()}} of \code{\link[=character]{character()}} vectors containing the levels of the columns in \code{data} specified by \code{varnames}
}
\description{
Extract levels from a set of variable names
}
\examples{
get_levels(iris, "Species")
}
