% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed_permutation_test_stat.R
\name{get_observed_permutation_test_stat}
\alias{get_observed_permutation_test_stat}
\title{Compute permutation test statistic}
\usage{
get_observed_permutation_test_stat(output_folder, dataset_names)
}
\arguments{
\item{output_folder}{where to find the relevant pickle files}

\item{dataset_names}{\code{\link[=character]{character()}} vector of dataset name stems}
}
\value{
the observed test statistic (a \code{\link[=numeric]{numeric()}} scalar)
}
\description{
Compute permutation test statistic
}
