% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_title_i_2.R
\name{get_title_i_2}
\alias{get_title_i_2}
\title{get title for PVD}
\usage{
get_title_i_2(subtype_and_stage_table, cval = FALSE, i)
}
\arguments{
\item{subtype_and_stage_table}{part of results \link[tibble:tbl_df-class]{tibble::tbl_df}}

\item{cval}{\link{logical}}

\item{i}{\link{integer}}
}
\value{
a \link{character} string
}
\description{
get title for PVD
}
\keyword{internal}
