% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gp34}
\alias{gp34}
\title{gp34}
\format{
## `gp34`
A data frame with nrow rows and ncol columns:
\describe{
 \item{col1}{col1name}
}
}
\source{
<url>
}
\usage{
gp34
}
\description{
Combined data from GP and GP4 studies.
}
\keyword{datasets}
