% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_CGG_repeats.R
\name{graph_CGG_repeats}
\alias{graph_CGG_repeats}
\title{Barplot of CGG repeats}
\usage{
graph_CGG_repeats(data)
}
\arguments{
\item{data}{a \link{data.frame} containing a \link{numeric} or \link{integer} column named \code{CGG}}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Barplot of CGG repeats
}
\examples{
visit1 |> graph_CGG_repeats()
}
