% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods_v2.R
\name{graph_likelihoods_v2}
\alias{graph_likelihoods_v2}
\title{Title}
\usage{
graph_likelihoods_v2(likelihoods, alpha = 0.5, facet = FALSE, legend = FALSE)
}
\arguments{
\item{likelihoods}{todo}

\item{alpha}{todo}

\item{facet}{whether to facet by number of subtypes}

\item{legend}{whether to show legend}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Title
}
\examples{
example_likelihoods |> graph_likelihoods_v2(alpha = 0.5)

}
