% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_abstract.R
\name{graphical_abstract}
\alias{graphical_abstract}
\title{Graphical Abstract}
\usage{
graphical_abstract(
  figs,
  tile_height = 1,
  y_text_size = 9,
  x_title_size = 8,
  facet_label_prefix = names(figs),
  legend.position = "none",
  scale_colors = c("red", "blue", "magenta", "darkgreen", "purple4"),
  rel_heights = c(1, 0.1),
  ncol_legend = 2,
  nrow_colors = 2,
  guide_rel_widths = c(0.7, 0.3),
  legend_text_size = grid::unit(8, "pt"),
  abstract,
  abstract_size = 5,
  abstract_width = 50,
  ...
)
}
\arguments{
\item{figs}{todo}

\item{tile_height}{todo}

\item{y_text_size}{todo}

\item{x_title_size}{todo}

\item{facet_label_prefix}{todo}

\item{legend.position}{todo}

\item{scale_colors}{todo}

\item{rel_heights}{(optional) Numerical vector of relative rows heights. Works just as
\code{rel_widths} does, but for rows rather than columns.}

\item{ncol_legend}{how many columns to use for legend}

\item{nrow_colors}{now many rows to use for color guide}

\item{guide_rel_widths}{relative widths of guides}

\item{legend_text_size}{\link[grid:unit]{grid::unit}: legend text size}

\item{abstract}{\link{character} String that summzrizes the study's main findings.}

\item{abstract_size}{\link{numeric} Abstract text size}

\item{abstract_width}{\link{numeric} Length at which to wrap the abstract text.}

\item{...}{List of plots to be arranged into the grid. The plots can be any objects that
the function \code{\link[cowplot:as_gtable]{as_gtable()}} can handle (see also examples).}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Graphical Abstract
}
