% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_colors.R
\name{group_colors}
\alias{group_colors}
\title{Extract group colors}
\usage{
group_colors(object)
}
\arguments{
\item{object}{an object}
}
\value{
a named \link{character} \link{vector} mapping
from \code{biomarker_group} to a color palette
}
\description{
Extract group colors
}
\examples{
figs <- readr::read_rds(
 testthat::test_path("fixtures", "sim-figs-4.rds")
)
}
