% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pySuStaIn.R
\name{install_pySuStaIn}
\alias{install_pySuStaIn}
\title{Install the \code{pySuStaIn} python package from github}
\usage{
install_pySuStaIn(envname = "r-pySuStaIn", python_version = "<3.10", ...)
}
\arguments{
\item{envname}{The name of, or path to, a conda environment.}

\item{python_version}{The version of Python to be installed. Set this if
you'd like to change the version of Python associated with a particular
conda environment.}

\item{...}{Optional arguments, reserved for future expansion.}
}
\description{
This function is a wrapper for \link[reticulate:conda-tools]{reticulate::conda_install}.
Its implementation follows the instructions in:
\itemize{
\item https://rstudio.github.io/reticulate/articles/package.html
\item https://rstudio.github.io/reticulate/articles/python_dependencies.html
}
}
