% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.list.R
\name{labels.list}
\alias{labels.list}
\title{Extract variable labels from list}
\usage{
\method{labels}{list}(object, ...)
}
\arguments{
\item{object}{a \link{list}}

\item{...}{unused}
}
\value{
a \link{vector} of \link{character} strings corresponding to \code{object}
}
\description{
Extract variable labels from list
}
