% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biomarker_events_table.R
\name{make_biomarker_events_table}
\alias{make_biomarker_events_table}
\title{Construct biomarker events table}
\usage{
make_biomarker_events_table(biomarker_levels, biomarker_groups)
}
\arguments{
\item{biomarker_levels}{todo}

\item{biomarker_groups}{todo}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
Construct biomarker events table
}
\examples{
biomarker_groups <- compile_biomarker_groups_table(
  dataset = test_data_v1
)

biomarker_varnames <-
  biomarker_groups |>
  dplyr::pull("biomarker")

biomarker_levels <-
  test_data_v1 |>
  get_levels(biomarker_varnames)

make_biomarker_events_table(
  biomarker_levels,
  biomarker_groups
)
}
