% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demographics_table.R
\name{make_demographics_table}
\alias{make_demographics_table}
\title{Make demographics table}
\usage{
make_demographics_table(
  data,
  strata = "Gender",
  vars = c("Age at visit", "Primary Race/Ethnicity", "FXTAS Stage", "CGG"),
  make_ft = TRUE,
  ...
)
}
\arguments{
\item{data}{a \link{data.frame} containing the variables specified by
\code{strata} and \code{vars}}

\item{strata}{names of column variable, specified as \link{character}}

\item{vars}{names of row variables, specified as \link{character}}

\item{make_ft}{\link{logical} whether to convert to flextable}

\item{...}{additional arguments passed to
\code{\link[=format_demographics_table_as_flextable]{format_demographics_table_as_flextable()}}}
}
\value{
a \link[flextable:flextable]{flextable::flextable}
}
\description{
Make demographics table
}
\examples{
test_data_v1 |> make_demographics_table()
}
