% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_codes_to_NA.R
\name{missing_codes_to_NA}
\alias{missing_codes_to_NA}
\title{Title}
\usage{
missing_codes_to_NA(x, NA_codes = c(777, 888, 999))
}
\arguments{
\item{x}{a numeric vector}

\item{NA_codes}{numeric values that should be turned into to NAs}
}
\description{
Title
}
