% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_pattern_by_ID.R
\name{missing_pattern_by_ID}
\alias{missing_pattern_by_ID}
\title{Title}
\usage{
missing_pattern_by_ID(data = gp34, variable = "ApoE")
}
\arguments{
\item{variable}{}
}
\description{
Title
}
\examples{
missing_patterns = missing_pattern_by_ID()
}
