% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_reasons.factor.R
\name{missingness_reasons.factor}
\alias{missingness_reasons.factor}
\title{Compute missingness reasons for a vector of factor variables}
\usage{
missingness_reasons.factor(x, x.clean = replace_missing_codes_with_NAs(x))
}
\arguments{
\item{x}{the original vector}

\item{x.clean}{the cleaned vector}
}
\value{
a vector containing the values of x corresponding to `NA`s in x.clean,
}
\description{
Compute missingness reasons for a vector of factor variables
}
