% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_reasons_numeric.R
\name{missingness_reasons.numeric}
\alias{missingness_reasons.numeric}
\title{Title}
\usage{
missingness_reasons.numeric(x, x.clean = clean_numeric(x, ...), ...)
}
\arguments{
\item{x}{the original vector to be turned into a numeric}

\item{x.clean}{the cleaned vector}

\item{...}{
  Arguments passed on to \code{\link{clean_numeric}}
  \describe{
    \item{\code{NA_codes}}{numeric values that should be turned into to NAs}
    \item{\code{extra_codes}}{codes besides 777, 888, 999 to turn into NAs; has no effect if the \code{NA_codes} argument is changed from default by the user.}
  }}
}
\value{
a \link{factor} \link{vector} corresponding to the elements of \code{x},
indicating the reasons why \code{x} was missing.
}
\description{
Title
}
