% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_correct.R
\name{pander.prob_correct}
\alias{pander.prob_correct}
\title{Print \code{prob_correct} objects as Pandoc markdown tables}
\usage{
\method{pander}{prob_correct}(x, ...)
}
\arguments{
\item{x}{a \code{prob_correct} object (from \code{\link[=compute_prob_correct]{compute_prob_correct()}})}

\item{...}{currently unused}
}
\description{
Print \code{prob_correct} objects as Pandoc markdown tables
}
\examples{
control_data <-
   test_data_v1 |>
     filter(`FX*` == "CGG <55") |>
     select(all_of(biomarker_varnames))
prob_correct =
control_data |>
  compute_prob_correct(
    max_prob = .95,
    biomarker_levels = biomarker_levels)
prob_correct |> pander()
}
