% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PF.R
\name{plot.PF}
\alias{plot.PF}
\title{Create positional variance diagram (PVD)}
\usage{
\method{plot}{PF}(PFs, size.y = 12, color_label = "Pr(stage)")
}
\arguments{
\item{color_label}{}
}
\value{
a "PVD" object (extends \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}})
}
\description{
Create positional variance diagram (PVD)
}
