% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compact_pvd_est.R
\name{plot_compact_pvd_est}
\alias{plot_compact_pvd_est}
\title{Plot compact PVD with point estimates only}
\usage{
plot_compact_pvd_est(
  figs,
  tile_height = 1,
  y_text_size = 9,
  facet_label_prefix = names(figs),
  legend.position = "none",
  scale_colors = c("red", "blue", "magenta", "darkgreen", "purple4")
)
}
\arguments{
\item{figs}{todo}

\item{tile_height}{todo}

\item{y_text_size}{\link{integer}: size of y-axis text}

\item{facet_label_prefix}{todo}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{scale_colors}{todo}
}
\description{
Plot compact PVD with point estimates only
}
