% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cv_loglik.R
\name{plot_cv_loglik}
\alias{plot_cv_loglik}
\title{Plot cross-validated per-fold log-likelihood distributions}
\usage{
plot_cv_loglik(cv_loglik)
}
\arguments{
\item{cv_loglik}{\code{\link[=matrix]{matrix()}} with rows corresponding to cross-validation folds and columsn corresponding to number of latent subgroups}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot cross-validated per-fold log-likelihood distributions
}
