% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CVIC.R
\name{plot_CVIC}
\alias{plot_CVIC}
\title{Plot a CVIC curve}
\usage{
plot_CVIC(CVIC)
}
\arguments{
\item{CVIC}{\code{\link[=numeric]{numeric()}} vector of CVIC values, ordered by number of subtypes}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot a CVIC curve
}
