% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Title}
\usage{
plot_heatmap(heatmap_data, label = "Pr(marker in this position)")
}
\arguments{
\item{heatmap_data}{tibble with columns "position", "name", and "value"}

\item{label}{Label for color scale}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Title
}
