% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_permutation_results.R
\name{plot_permutation_results}
\alias{plot_permutation_results}
\title{Plot permutation test results}
\usage{
plot_permutation_results(permuted_test_stats, observed_test_stat)
}
\arguments{
\item{permuted_test_stats}{a \code{\link[=numeric]{numeric()}} vector}

\item{observed_test_stat}{a \code{\link[=numeric]{numeric()}} with length 1}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
\description{
Plot permutation test results
}
