% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positional_variance.R
\name{plot_positional_var}
\alias{plot_positional_var}
\title{Title}
\usage{
plot_positional_var(
  samples_sequence,
  samples_f,
  n_samples,
  score_vals,
  biomarker_labels = NULL,
  ml_f_EM = NULL,
  cval = FALSE,
  subtype_order = NULL,
  biomarker_order = NULL,
  title_font_size = 12,
  stage_font_size = 10,
  stage_label = "SuStaIn Stage",
  stage_rot = 0,
  stage_interval = 1,
  label_font_size = 10,
  label_rot = 0,
  cmap = "original",
  biomarker_colours = NULL,
  figsize = NULL,
  subtype_titles = NULL,
  separate_subtypes = FALSE,
  save_path = NULL,
  save_kwargs = NULL
)
}
\arguments{
\item{save_kwargs}{}
}
\description{
Title
}
