% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{print.permutation_test}
\alias{print.permutation_test}
\title{print method for "permutation_test" objects}
\usage{
\method{print}{permutation_test}(x, ...)
}
\arguments{
\item{x}{a \code{\link[=permutation_test]{permutation_test()}} object}

\item{...}{unused}
}
\value{
\code{x}, invisibly
}
\description{
print method for "permutation_test" objects
}
