% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_PVDs.R
\name{print_PVDs}
\alias{print_PVDs}
\title{Print a list of PVDs}
\usage{
print_PVDs(figs)
}
\arguments{
\item{figs}{a \link{list} of \link[ggplot2:ggplot]{ggplot2::ggplot} objects}
}
\value{
\code{figs}, invisibly
}
\description{
Print a list of PVDs
}
