% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd_lineplot.R
\name{pvd_lineplot}
\alias{pvd_lineplot}
\title{Plot change in Stage ranking}
\usage{
pvd_lineplot(
  figs,
  events_to_highlight = NULL,
  highlight_color = "yellow",
  min_alpha = 0.5,
  max_alpha = 1,
  stage_alpha = 1,
  facet_labels = names(figs),
  text_size = 3.4,
  y_lab = "Sequential order",
  y_title_size = 9,
  y_text_size = 8,
  x_text_size = grid::unit(8, "points"),
  nrow_group_col = 1,
  x_positions = c(1.01, 1.14),
  direction_colors = c(`(stage)` = "grey25", down = "#DC3220", neutral = "grey70", up =
    "#005AB5"),
  expand = 0,
  group_cols = group_colors(figs),
  rel_heights = c(6, 1),
  legend_text_size = grid::unit(8, "pt"),
  guide_group_colors = TRUE
)
}
\arguments{
\item{figs}{a \link{list} of todo}

\item{events_to_highlight}{todo}

\item{highlight_color}{todo}

\item{min_alpha}{todo}

\item{max_alpha}{todo}

\item{stage_alpha}{todo}

\item{facet_labels}{todo}

\item{text_size}{todo}

\item{y_lab}{todo}

\item{y_title_size}{todo}

\item{y_text_size}{todo}

\item{x_text_size}{todo}

\item{nrow_group_col}{\link{integer} how many rows in group_col legend}

\item{x_positions}{positions of x-axis ticks}

\item{direction_colors}{a \link{character} vector of length 4}

\item{expand}{\link{numeric} how much to pad the sides}

\item{group_cols}{a named \link{character} \link{vector}
mapping from \code{biomarker_group} to a color palette}

\item{rel_heights}{(optional) Numerical vector of relative rows heights. Works just as
\code{rel_widths} does, but for rows rather than columns.}

\item{legend_text_size}{\link[grid:unit]{grid::unit}: legend text size}

\item{guide_group_colors}{\link{logical}:
whether to provide a guide for group colors}
}
\description{
Plot change in Stage ranking
}
\examples{
output_path <-
  fs::path_package("extdata/sim_data", package = "fxtas")

figs <- extract_figs_from_pickle(
  output_folder = output_path,
  use_rds = TRUE,
  n = 3
)

figs[1:2] |> pvd_lineplot(
  group_cols = group_colors(figs)
)
}
