% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd_subtype_lineplot.R
\name{pvd_subtype_lineplot}
\alias{pvd_subtype_lineplot}
\title{Display Sequential Order of Events by Subtype}
\usage{
pvd_subtype_lineplot(
  figs,
  events_to_highlight,
  min_alpha = 0.25,
  max_alpha = 1,
  stage_alpha = 1,
  facet_names = names(figs),
  facet_label_prefixes = NULL,
  facet_labels = unlist(compact_pvd_facet_labels(figs = figs, facet_label_prefix =
    facet_label_prefixes)),
  text_size = 3.4,
  y_lab = "Sequential order",
  y_title_size = 9,
  y_text_size = 8,
  x_text_size = 8,
  subtype_x = c(1, 3, 5, 7),
  mult = 0.2,
  align_stage = TRUE,
  use_group_colors = FALSE,
  group_colors = group_colors(figs),
  ...
)
}
\arguments{
\item{figs}{A \link{list} of todo}

\item{events_to_highlight}{todo}

\item{min_alpha}{todo}

\item{max_alpha}{todo}

\item{stage_alpha}{todo}

\item{facet_names}{\link{character}: facet names}

\item{facet_label_prefixes}{\link{character}: prefixes for facet labels}

\item{facet_labels}{todo}

\item{text_size}{todo}

\item{y_lab}{todo}

\item{y_title_size}{todo}

\item{y_text_size}{todo}

\item{x_text_size}{todo}

\item{subtype_x}{Vector of x-axis value for the subtypes.}

\item{mult}{\link{numeric} vector}

\item{align_stage}{\link{logical} whether to align by FXTAS stage}

\item{use_group_colors}{\link{logical}: whether to use existing group colors
or custom line colors from \code{events_to_highlight} argument}

\item{group_colors}{a named \link{character} \link{vector} mapping from \code{biomarker_group} to a color palette}

\item{...}{additional arguments passed to \link[ggbump:geom_bump]{ggbump::geom_bump}
Default = c(1, 1.15, 1.75, 2.35)}
}
\description{
Display Sequential Order of Events by Subtype
}
