% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missingness_reasons_vars.R
\name{remove_missingness_reason_vars}
\alias{remove_missingness_reason_vars}
\title{Remove variable}
\usage{
remove_missingness_reason_vars(vars)
}
\arguments{
\item{vars}{a \code{\link[=character]{character()}} vector}
}
\value{
a modified version of input \code{vars} (\code{\link[=character]{character()}}) with elements ending in "missingness reasons" removed
}
\description{
Remove variable
}
\examples{
vars = c(
"parkinsonian features",
"parkinsonian features: missingness reasons")
vars |> remove_missingness_reason_vars()
}
