% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_codes.R
\name{replace_missing_codes}
\alias{replace_missing_codes}
\title{Replace missing codes in a vector of factor variables}
\usage{
replace_missing_codes(
  x,
  missing_codes = c("99", "777", "888", "999",
    "Asked by clinician, but no answer from subject (99)",
    "Question not asked at time of data entry; check records (777)", "NA (888)",
    "No Response (999)", "Missing/Refused (999)"),
  replacement = levels(x)[1],
  mapping = setNames(rep(replacement, length(missing_codes)), missing_codes)
)
}
\arguments{
\item{x}{a vector of factor variables}

\item{missing_codes}{a vector of strings}

\item{replacement}{a string variable or \code{NA}}

\item{mapping}{a named string vector}
}
\description{
The default is to use the reference level of the factor.
}
\examples{
x = iris$Species |> unique()
x |> replace_missing_codes(missing_codes = "virginica", replacement = "versicolor")
}
