% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_codes_with_NAs.R
\name{replace_missing_codes_with_NAs}
\alias{replace_missing_codes_with_NAs}
\title{Replace missing codes with NAs}
\usage{
replace_missing_codes_with_NAs(x)
}
\arguments{
\item{x}{a factor or character vector}
}
\value{
a factor vector similar to `x`, but with 99, 777, 888, 999, etc replaced with `NA`s
}
\description{
Replace missing codes with NAs
}
