% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sex_differences.R
\name{report_sex_differences}
\alias{report_sex_differences}
\title{Report sex differences}
\usage{
report_sex_differences(table)
}
\arguments{
\item{table}{output from \code{\link[=make_biomarkers_table]{make_biomarkers_table()}}}
}
\value{
a \code{\link[=character]{character()}} string
}
\description{
Report sex differences
}
