% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_s3_sim_data.R
\docType{data}
\name{results_s3_sim_data}
\alias{results_s3_sim_data}
\title{simulated SuStaIn results pickle data}
\format{
An object of class \code{SuStaIn_model} (inherits from \code{list}) of length 15.
}
\source{
\code{SuStaIn-simulated-data.qmd}
}
\usage{
results_s3_sim_data
}
\description{
A \code{"SuStaIn_model"} object (from \code{\link[=extract_results_from_pickle]{extract_results_from_pickle()}})
}
\keyword{datasets}
