% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSuStaIn-package.R
\docType{package}
\name{rSuStaIn-package}
\alias{rSuStaIn}
\alias{rSuStaIn-package}
\title{rSuStaIn: R Implementation of SuStaIn Algorithm}
\description{
An R package for applying the Ordinal SuStaIn (Subtype and Stage Inference) algorithm to study disease progression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ucd-iddrc/rSuStaIn}
  \item \url{https://ucd-iddrc.github.io/rSuStaIn/}
  \item Report bugs at \url{https://github.com/ucd-iddrc/rSuStaIn/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu}

Authors:
\itemize{
  \item Matthew Ponzini \email{mdponzini@ucdavis.edu}
}

}
\keyword{internal}
