% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SSA.R
\name{run_SSA}
\alias{run_SSA}
\title{Title}
\usage{
run_SSA(
  prob_score,
  score_vals,
  SuStaInLabels,
  N_startpoints,
  N_S_max,
  N_iterations_MCMC,
  output_folder,
  dataset_name,
  use_parallel_startpoints = FALSE,
  seed = 1,
  plot = FALSE
)
}
\arguments{
\item{plot}{}
}
\description{
Title
}
