% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scid_vars_for_table1.R
\name{scid_vars_for_table1}
\alias{scid_vars_for_table1}
\title{List SCID variables to display in data summary tables}
\usage{
scid_vars_for_table1()
}
\value{
a \code{\link[=character]{character()}} vector
}
\description{
List SCID variables to display in data summary tables
}
\examples{
scid_vars_for_table1()
}
