% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_articles.R
\name{search_articles}
\alias{search_articles}
\title{Search PubMed for FXTAS-related articles}
\usage{
search_articles(query)
}
\arguments{
\item{query}{a \link{character} string}
}
\value{
a \link[tibble:tbl_df]{tibble::tbl_df}
}
\description{
Search PubMed for FXTAS-related articles
}
\examples{
\dontrun{
search_articles("FXTAS")
}
}
