% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_data}
\alias{sim_data}
\title{Simulated data}
\format{
\subsection{\code{sim_data}}{

A data frame with nrow rows and ncol columns:
\describe{
\item{col1}{col1name}
}
}
}
\source{
\if{html}{\out{
<url>
}}
}
\usage{
sim_data
}
\description{
Toy example data provided by Alexandra Young
}
\keyword{datasets}
