% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_barplot.R
\name{stage_barplot.default}
\alias{stage_barplot.default}
\title{Title}
\usage{
\method{stage_barplot}{default}(object, include_type_0 = TRUE, ...)
}
\arguments{
\item{include_type_0}{}

\item{...}{unused}
}
\description{
Title
}
