% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_barplot.R
\name{stage_barplot.list}
\alias{stage_barplot.list}
\title{Plot estimated stage counts}
\usage{
\method{stage_barplot}{list}(object)
}
\arguments{
\item{results_list}{a \link{list} of \code{Sustain_model} objects}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot estimated stage counts
}
