% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_formula.R
\name{stratified_formula}
\alias{stratified_formula}
\title{Build a formula for \code{\link[table1:table1]{table1::table1()}}}
\usage{
stratified_formula(vars, strata)
}
\arguments{
\item{vars}{row variables}

\item{strata}{column variables}
}
\value{
a \code{\link[=formula]{formula()}} object
}
\description{
Build a formula for \code{\link[table1:table1]{table1::table1()}}
}
\examples{
stratified_formula(c("Sepal.Length" ,"Sepal.Width"), "Species")
}
