% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.permutation_test.R
\name{summary.permutation_test}
\alias{summary.permutation_test}
\title{Summary method for \code{permutation_test} objects}
\usage{
\method{summary}{permutation_test}(object, N, digits = 2, ...)
}
\arguments{
\item{object}{a \code{permutation_test} object (from \code{\link[=permutation_test]{permutation_test()}})}

\item{N}{sample size of underlying data}

\item{digits}{how many digits to round the test statistic}

\item{...}{unused}
}
\value{
a \link[glue:glue]{glue::glue} string
}
\description{
Summary method for \code{permutation_test} objects
}
\examples{
\dontrun{
output_folder <-
  "output/output.fixed_CV-scid-no-subthres/"

permuted_test_stats_sex <- collect_permutation_test_stats(
  output_folder = output_folder,
  permuting_variables = "Gender"
)

observed_test_stat_sex <- get_observed_permutation_test_stat(
  output_folder,
  dataset_names = c("females", "males")
)

pval_sex_gp34_v1 <- permutation_test(
  observed_test_stat = observed_test_stat_sex,
  permuted_test_stats = permuted_test_stats_sex,
  n_permutations = n_permutations
)

summary(pval_sex_gp34_v1, N = 253)
}
}
