% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1a.R
\name{table1a}
\alias{table1a}
\title{\code{\link[table1:table1]{table1::table1()}} with customized settings}
\usage{
table1a(...)
}
\arguments{
\item{...}{arguments passed to \code{\link[table1:table1]{table1::table1()}}}
}
\value{
an object of class \code{"table1"}
}
\description{
\code{\link[table1:table1]{table1::table1()}} with customized settings
}
