% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_subtype_by_gender.R
\name{table_subtype_by_gender}
\alias{table_subtype_by_gender}
\title{Title}
\usage{
table_subtype_by_gender(patient_data, subtype_and_stage_table)
}
\arguments{
\item{subtype_and_stage_table}{}
}
\description{
Title
}
