% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_label.R
\name{var_label.list}
\alias{var_label.list}
\title{Extract variable labels from a list object}
\usage{
\method{var_label}{list}(x, ...)
}
\arguments{
\item{x}{a \link{list}}

\item{...}{not used}
}
\value{
a named \link{character} \link{vector}
}
\description{
Extract variable labels from a list object
}
\examples{
output_folder <-
  fs::path_package("extdata/sim_data/", package = "fxtas")

biomarker_levels =
  output_folder |>
  fs::path("biomarker_levels.rds") |>
  readr::read_rds()

biomarker_levels |> var_label()
}
