% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_permuted_test_stats.R
\name{write_permuted_test_stats}
\alias{write_permuted_test_stats}
\title{Write permuted test stats as RDS}
\usage{
write_permuted_test_stats(
  permuted_test_stats,
  permutations = names(permuted_test_stats),
  output_folder,
  verbose = TRUE
)
}
\arguments{
\item{permuted_test_stats}{a named \code{\link[=numeric]{numeric()}} vector of permuted test statistics}

\item{permutations}{the vector of permutation seeds used to generate \code{permuted_test_stats}}

\item{output_folder}{where to create a \code{test_stats} folder and save \code{permuted_test_stats} as an .rds file}

\item{verbose}{whether to print informative messages about progress}
}
\description{
Write permuted test stats as RDS
}
