% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_potential_biomarkers.R
\name{all_potential_biomarkers}
\alias{all_potential_biomarkers}
\title{Title}
\usage{
all_potential_biomarkers(dataset = gp34)
}
\arguments{
\item{dataset}{Dataset to extract names from}
}
\description{
Title
}
