% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_numeric.R
\name{clean_numeric}
\alias{clean_numeric}
\title{Title}
\usage{
clean_numeric(x, NA_codes = c(777, 888, 999, extra_codes), extra_codes = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{NA_codes}{numeric values that should be turned into to NAs}

\item{extra_codes}{codes besides 777, 888, 999 to turn into NAs}
}
\description{
Title
}
