% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_numeric.R
\name{clean_numeric}
\alias{clean_numeric}
\title{Clean a Numeric Variable}
\usage{
clean_numeric(x, NA_codes = c(777, 888, 999, extra_codes), extra_codes = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{NA_codes}{numeric values that should be turned into to NAs}

\item{extra_codes}{codes besides 777, 888, 999 to turn into NAs; has no effect if the `NA_codes` argument is changed from default by the user.}
}
\description{
Replace missing codes (e.g., 777, 888, 999) with NAs in a numeric variable
}
