% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_permutation_test_stats.R
\name{collect_permutation_test_stats}
\alias{collect_permutation_test_stats}
\title{Collect permutation test statistics}
\usage{
collect_permutation_test_stats(
  permuting_variables,
  stratifying_levels = "",
  output_folder = fs::path(here::here(), "output/output.fixed_CV"),
  permutation_results_folder_path = fs::path(output_folder, "permutations",
    paste(fs::path_sanitize(stratifying_levels), collapse = "-"),
    paste(fs::path_sanitize(permuting_variables), collapse = "-"), "test_stats"),
  file_stem = "permuted_test_stats",
  seeds = 1:1000,
  by = 20,
  first_seeds = seq(min(seeds), max(seeds), by = by),
  last_seeds = first_seeds + by - 1
)
}
\arguments{
\item{output_folder}{where to find the \code{permutations/test_stats}
subfolder}

\item{permutation_results_folder_path}{path to folder
containing \code{.rds} files with permuted test statistics}

\item{file_stem}{stem of test stat .rds files}

\item{seeds}{seeds to find}

\item{by}{number of seeds per file}
}
\value{
a \link{numeric} vector
}
\description{
Collect permutation test statistics
}
\examples{
\dontrun{
test_stats = collect_permutation_test_stats("output/output.fixed_CV")
}
}
