% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_pvd_figure.R
\name{compact_pvd_figure}
\alias{compact_pvd_figure}
\title{plot compact pvd: figure}
\usage{
compact_pvd_figure(
  plot_dataset,
  tile_height,
  y_text_size,
  facet_names,
  legend.position,
  scale_colors
)
}
\description{
plot compact pvd: figure
}
