% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_biomarker_group_list.R
\name{compile_biomarker_group_list}
\alias{compile_biomarker_group_list}
\title{Compile list of biomarker groups}
\usage{
compile_biomarker_group_list(dataset = gp34)
}
\arguments{
\item{dataset}{Dataset to extract biomarker names from using regular expressions}
}
\value{
a \link{list} of \link{character} vectors, with the list names denoting biomarker groups and the elements of the \link{character} vectors denoting individual biomarkers in each group.
}
\description{
This function compiles a \link{list} of biomarker variable names, grouped into categories.
Some of the categories are hard-coded, while others are extracted from the
column names of the input \link{data.frame} \code{dataset} using regular expressions (see \link[base:regex]{base::regex} and \code{\link[base:grep]{base::grep()}}).
}
\examples{
gp34 |> compile_biomarker_group_list()

}
