% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_biomarker_groups_table.R
\name{compile_biomarker_groups_table}
\alias{compile_biomarker_groups_table}
\title{Compile biomarker groups table}
\usage{
compile_biomarker_groups_table(
  biomarker_group_list = compile_biomarker_group_list(...),
  colors = choose_biomarker_group_colors(biomarker_group_list),
  ...
)
}
\arguments{
\item{biomarker_group_list}{todo}
}
\value{
a \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Compile biomarker groups table
}
