% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_confus_matrix.R
\name{compute_confus_matrix}
\alias{compute_confus_matrix}
\title{compute \code{confus_matrix} as in python version}
\usage{
compute_confus_matrix(
  samples_sequence,
  biomarker_event_order = attr(samples_sequence, "biomarker_event_names")
)
}
\arguments{
\item{samples_sequence}{todo}

\item{biomarker_event_order}{todo}
}
\value{
a \link{tibble:tbl_df}
}
\description{
compute \code{confus_matrix} as in python version
}
\examples{
samples_sequence = matrix(
compute_confus_matrix(samples_sequence)

}
