% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_confus_matrix.R
\name{compute_confus_matrix}
\alias{compute_confus_matrix}
\title{Title}
\usage{
compute_confus_matrix(samples_sequence)
}
\arguments{
\item{samples_sequence}{}
}
\description{
Title
}
\examples{
samples_sequence = matrix(
compute_confus_matrix(samples_sequence)
}
