% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_correct.R
\name{compute_prob_correct}
\alias{compute_prob_correct}
\title{Title}
\usage{
compute_prob_correct(dataset, biomarkers, DataScores)
}
\arguments{
\item{dataset}{Should be just the control data?}

\item{DataScores}{}
}
\description{
Title
}
