% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_correct.R
\name{compute_prob_correct}
\alias{compute_prob_correct}
\title{Title}
\usage{
compute_prob_correct(
  dataset,
  biomarkers = names(biomarker_levels),
  biomarker_levels,
  DataScores,
  max_prob = 1
)
}
\arguments{
\item{dataset}{Should be just the control data?}

\item{biomarker_levels}{a list containing the levels of each biomarker}

\item{max_prob}{}
}
\description{
Title
}
