% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prob_scores.R
\name{compute_prob_scores}
\alias{compute_prob_scores}
\title{Title}
\usage{
compute_prob_scores(
  dataset,
  biomarker_varnames,
  ModelScores,
  DataScores,
  prob_dist = compute_prob_dist(ModelScores = ModelScores, DataScores = DataScores,
    biomarkers = biomarker_varnames, ...),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{prob_dist}{}
}
\description{
Title
}
