% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scid_domains.R
\name{create_scid_domains}
\alias{create_scid_domains}
\title{Create SCID domain variables}
\usage{
create_scid_domains(
  dataset,
  scid_vars = c("Bipolar I Disorder (MD01), Lifetime",
    "Bipolar I Disorder (MD01), Current", "Bipolar II Disorder (MD02), Lifetime",
    "Bipolar II Disorder (MD02), Current", "Other Bipolar Disorder (MD03), Current",
    "Other Bipolar Disorder (MD03), Lifetime",
    "Major Depressive Disorder (MD04), Current",
    "Major Depressive Disorder (MD04), Lifetime",
    "Mood Disorder Due to GMC (MD07), Current",
    "Mood Disorder Due to GMC (MD07), Lifetime",
    "Substance-Induced Mood Dis. (MD08), Current",
    "Substance-Induced Mood Dis. (MD08), Lifetime", 
    
    "Primary Psychotic Symptoms (PS01), Current",
    "Primary Psychotic Symptoms (PS01), Lifetime", "Dysthymic Disorder (MD05), Lifetime",
    "Depressive Disorder NOS (MD06), Lifetime", "Depressive Disorder NOS (MD06) Current",
    "Alcohol (SUD17), Lifetime", "Alcohol (SUD17) Current",
    "Sedative-Hypnotic-Anxiolytic (SUD18), Lifetime",
    "Sedative-Hypnotic-Anxiolytic (SUD18), Current", "Cannabis (SUD19), Lifetime",
    "Cannabis (SUD19),Current", "Stimulants (SUD20), Lifetime",
    "Stimulants (SUD20) Current", "Opiod (SUD21), Lifetime", 
    
    "Cocaine (SUD22), Lifetime", "Opiod (SUD21), Current",
    "Hallucinogenics/ PCP (SUD23), Lifetime", "Cocaine (SUD22) Current",
    "Hallucinogenics/ PCP (SUD23), Current", "Poly Drug (SUD24), Lifetime",
    "Poly Drug (SUD24), Current", "Other (SUD25), Lifetime", "Other (SUD25), Current",
    "Panic Disorder (ANX26), Lifetime", "Panic Disorder (ANX26), Current",
    "Agoraphobia without Panic (ANX27), Lifetime",
    "Agoraphobia without Panic (ANX27), Current", "Social Phobia (ANX28), Lifetime",
    "Social Phobia (ANX28), Current", 
     "Specific Phobia (ANX29), Lifetime",
    "Specific Phobia (ANX29), Current", "Obsessive Compulsive (ANX30), Lifetime",
    "Obsessive Compulsive (ANX30), Current", "Posttraumatic Stress (ANX31), Lifetime",
    "Posttraumatic Stress (ANX31), Current", "Generalized Anxiety (ANX32), Current Only",
    "Anxiety Due to GMC (ANX33), Lifetime", "Anxiety Due to GMC (ANX33), Current",
    "Substance-Induced Anxiety (ANX34), Lifetime",
    "Substance-Induced Anxiety (ANX34), Current",
    "Anxiety Disorder NOS (ANX35), Lifetime", 
    
    "Anxiety Disorder NOS (ANX35), Current", "Somatization Disorder (SOM36)",
    "Pain Disorder (SOM37)", "Undifferentiated Somatoform (SOM38)",
    "Body Dysmorphic (SOM40)", "Hypochondriasis (SOM39)")
)
}
\arguments{
\item{dataset}{a \link[tibble:tbl_df-class]{tibble::tbl_df} containing all of \code{scid_vars} as columns}

\item{scid_vars}{a \code{\link[=character]{character()}} vector}
}
\value{
a \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Create SCID domain variables
}
