% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvic_sim_data.R
\docType{data}
\name{cvic_sim_data}
\alias{cvic_sim_data}
\title{CVIC results from simulated data}
\format{
An object of class \code{CVIC} (inherits from \code{list}) of length 2.
}
\usage{
cvic_sim_data
}
\description{
generated by \code{\link[=extract_cvic]{extract_cvic()}};
see \code{SuStaIn-simulated-data.qmd} for the full script to generate this data
}
\keyword{datasets}
