% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_figs_from_pickle.R
\name{extract_figs_from_pickle}
\alias{extract_figs_from_pickle}
\title{Title}
\usage{
extract_figs_from_pickle(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  score_vals,
  SuStaInLabels,
  picklename = paste0(dataset_name, "_subtype", n_s - 1, ".pickle"),
  results = py_load_object(fs::path(output_folder, "pickle_files", picklename)),
  ...
)
}
\arguments{
\item{results}{}
}
\description{
Title
}
