% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_figs_from_pickle.R
\name{extract_figs_from_pickle}
\alias{extract_figs_from_pickle}
\title{Extact PVDs from pickle file}
\usage{
extract_figs_from_pickle(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  rda_filename = "data.RData",
  picklename = paste0(dataset_name, "_subtype", n_s - 1, ".pickle"),
  use_rds = TRUE,
  ...
)
}
\arguments{
\item{n_s}{helps construct \code{picklename} by identifying number of latent subgroups}

\item{dataset_name}{root name of dataset}

\item{output_folder}{where to find the dataset}

\item{rda_filename}{name of rda file containing environment used to run analyses}

\item{picklename}{the name of the pickle file to open}

\item{...}{
  Arguments passed on to \code{\link[=plot_positional_var]{plot_positional_var}}
  \describe{
    \item{\code{samples_sequence}}{todo}
    \item{\code{samples_f}}{todo}
    \item{\code{n_samples}}{todo}
    \item{\code{score_vals}}{todo}
    \item{\code{biomarker_labels}}{todo}
    \item{\code{ml_f_EM}}{todo}
    \item{\code{cval}}{todo}
    \item{\code{subtype_order}}{todo}
    \item{\code{biomarker_order}}{todo}
    \item{\code{title_font_size}}{todo}
    \item{\code{stage_font_size}}{todo}
    \item{\code{stage_label}}{todo}
    \item{\code{stage_rot}}{todo}
    \item{\code{stage_interval}}{todo}
    \item{\code{label_font_size}}{todo}
    \item{\code{label_rot}}{todo}
    \item{\code{cmap}}{a \link{character}}
    \item{\code{biomarker_colours}}{a \link{character} vector of colors}
    \item{\code{subtype_titles}}{todo}
    \item{\code{separate_subtypes}}{todo}
    \item{\code{save_path}}{todo}
    \item{\code{save_kwargs}}{todo}
    \item{\code{results}}{todo}
    \item{\code{biomarker_levels}}{todo}
    \item{\code{biomarker_groups}}{biomarker groupings}
    \item{\code{biomarker_events_table}}{todo}
    \item{\code{biomarker_event_names}}{todo}
    \item{\code{biomarker_plot_order}}{todo}
    \item{\code{synchronize_y_axes}}{todo}
  }}
}
\value{
a \verb{"PVD.list} (a \link{list} of \code{PVD} objects from \code{\link[=plot.PF]{plot.PF()}})
}
\description{
Extact PVDs from pickle file
}
\examples{

output_path =
  fs::path_package("extdata/sim_data", package = "fxtas")

if(dir.exists(output_path))
{
figs = extract_figs_from_pickle(
  output_folder = output_path,
  n = 3)

 figs
}
}
