% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results_from_pickle.R
\name{extract_results_from_pickle}
\alias{extract_results_from_pickle}
\title{Extract results from pickle file}
\usage{
extract_results_from_pickle(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  rda_filename = "data.RData",
  basename = paste0(dataset_name, "_subtype", n_s - 1),
  picklename = paste0(basename, ".pickle"),
  format_results = TRUE,
  use_rds = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_s}{helps construct \code{picklename} by identifying number of latent subgroups}

\item{dataset_name}{root name of dataset}

\item{output_folder}{where to find the dataset}

\item{rda_filename}{name of rda file containing environment used to run analyses}

\item{picklename}{the name of the pickle file to open}

\item{format_results}{whether to apply \code{\link[=format_results_list]{format_results_list()}} to results before returning}

\item{verbose}{whether to print messages}

\item{...}{
  Arguments passed on to \code{\link[=format_results_list]{format_results_list}}
  \describe{
    \item{\code{format_sst}}{should the subtype and stage table be formatted? (doesn't work for cross-validation fold pickle-files)}
  }}
}
\value{
a \code{"SuStaIn-model"} object (extends \code{\link[=list]{list()}})
}
\description{
Extract results from pickle file
}
\examples{
output_path =
  fs::path_package("extdata/sim_data", package = "fxtas")
results = extract_results_from_pickle(output_folder = output_path)
}
