% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results_from_pickle.R
\name{extract_results_from_pickle}
\alias{extract_results_from_pickle}
\title{Title}
\usage{
extract_results_from_pickle(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  picklename = paste0(dataset_name, "_subtype", n_s - 1, ".pickle"),
  results = format_results_list(py_load_object(fs::path(output_folder, "pickle_files",
    picklename)), ...),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
