% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results_from_pickle.R
\name{extract_results_from_pickles}
\alias{extract_results_from_pickles}
\title{Extract results from multiple pickle files}
\usage{
extract_results_from_pickles(
  n_s = 1,
  dataset_name = "sample_data",
  output_folder = "output",
  rda_filename = "data.RData",
  basename = paste0(dataset_name, "_subtype", n_s - 1),
  picklename = paste0(basename, ".pickle"),
  format_results = TRUE,
  use_rds = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{n_s}{helps construct \code{picklename} by identifying number of latent subgroups}

\item{dataset_name}{root name of dataset}

\item{output_folder}{where to find the dataset}

\item{rda_filename}{name of rda file containing environment used to run analyses}

\item{picklename}{the name of the pickle file to open}

\item{format_results}{whether to apply \code{\link[=format_results_list]{format_results_list()}} to results before returning}

\item{verbose}{whether to print messages}
}
\value{
a \link{list} of \code{"SuStaIn-model"} objects (extends \code{\link[=list]{list()}})
}
\description{
Extract results from multiple pickle files
}
