% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_biomarkers_table.R
\name{flex_biomarkers_table}
\alias{flex_biomarkers_table}
\title{Print method for \code{biomarkers_table}}
\usage{
flex_biomarkers_table(x)
}
\arguments{
\item{x}{a "biomarkers_table" object from \link{make_biomarkers_table(0)}}
}
\value{
a \code{\link[flextable:flextable]{flextable::flextable()}}
}
\description{
Print method for \code{biomarkers_table}
}
