% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_demographics_table_as_flextable.R
\name{format_demographics_table_as_flextable}
\alias{format_demographics_table_as_flextable}
\title{Title}
\usage{
format_demographics_table_as_flextable(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\keyword{internal}
