% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_events_table.R
\name{get_biomarker_events_table}
\alias{get_biomarker_events_table}
\title{Title}
\usage{
get_biomarker_events_table(biomarker_levels)
}
\arguments{
\item{biomarker_levels}{todo}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Title
}
\examples{
df = test_data_v1
biomarker_groups = compile_biomarker_groups_table(dataset = df)
biomarker_varnames =
  biomarker_groups |>
  pull("biomarker")

biomarker_levels =
  df |>
  get_levels(biomarker_varnames, keep_labels = TRUE)

 get_biomarker_events_table(biomarker_levels)
}
