% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomarker_events_table.R
\name{get_biomarker_events_table}
\alias{get_biomarker_events_table}
\title{Title}
\usage{
get_biomarker_events_table(biomarker_levels)
}
\arguments{
\item{biomarker_levels}{todo}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Title
}
