% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_title_i_2.R
\name{get_title_i_2}
\alias{get_title_i_2}
\title{get title for PVD}
\usage{
get_title_i_2(subtype_and_stage_table, cval = FALSE, i)
}
\arguments{
\item{subtype_and_stage_table}{part of results \link[tibble:tbl_df-class]{tibble::tbl_df}}

\item{cval}{\link{logical}}

\item{i}{\link{integer}}
}
\value{
a \link{character} string
}
\description{
get title for PVD
}
\examples{
sim_results_3s =
 fs::path_package("extdata/sim_data/", package = "fxtas") |>
  extract_results_from_pickle(n_s = 3,
                              output_folder = _,
                              use_rds = FALSE)
 sustain_table = sim_results_3s$subtype_and_stage_table
 get_title_i_2(sustain_table, cval = FALSE, i = 1)
}
\keyword{internal}
