% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_CGG_repeats.R
\name{graph_CGG_repeats}
\alias{graph_CGG_repeats}
\title{Barplot of CGG repeats}
\usage{
graph_CGG_repeats(data, rows = NULL, cols = NULL, include_200_cutoff = FALSE)
}
\arguments{
\item{data}{a \link{data.frame} containing a \link{numeric} or \link{integer} column named \code{CGG}}

\item{rows, cols}{A set of variables or expressions quoted by
\code{\link[ggplot2:vars]{vars()}} and defining faceting groups on the rows or columns
dimension. The variables can be named (the names are passed to
\code{labeller}).

For compatibility with the classic interface, \code{rows} can also be
a formula with the rows (of the tabular display) on the LHS and
the columns (of the tabular display) on the RHS; the dot in the
formula is used to indicate there should be no faceting on this
dimension (either row or column).}

\item{include_200_cutoff}{whether to include the \code{CGG = 200} cutoff}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot} object
}
\description{
Barplot of CGG repeats
}
\examples{
visit1 |> graph_CGG_repeats()
visit1 |> graph_CGG_repeats(cols = vars(Gender))
}
