% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods_v2.R
\name{graph_likelihoods_v2}
\alias{graph_likelihoods_v2}
\title{Title}
\usage{
graph_likelihoods_v2(likelihoods, alpha = 0.5)
}
\arguments{
\item{likelihoods}{todo}

\item{alpha}{todo}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Title
}
