% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pySuStaIn.R
\name{install_pySuStaIn}
\alias{install_pySuStaIn}
\title{Install the \code{pySuStaIn} python package from github}
\usage{
install_pySuStaIn(envname = "r-pySuStaIn", method = "auto", ...)
}
\arguments{
\item{envname}{The name, or full path, of the environment in which Python
packages are to be installed. When \code{NULL} (the default), the active
environment as set by the \code{RETICULATE_PYTHON_ENV} variable will be used;
if that is unset, then the \code{r-reticulate} environment will be used.}

\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{...}{Additional arguments passed to \code{\link[reticulate:conda_install]{conda_install()}}
or \code{\link[reticulate:virtualenv_install]{virtualenv_install()}}.}
}
\description{
Install the \code{pySuStaIn} python package from github
}
