% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biomarker_events_table.R
\name{make_biomarker_events_table}
\alias{make_biomarker_events_table}
\title{Construct biomarker events table}
\usage{
make_biomarker_events_table(biomarker_levels, biomarker_groups)
}
\arguments{
\item{biomarker_levels}{todo}

\item{biomarker_groups}{todo}
}
\value{
a \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Construct biomarker events table
}
