% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_biomarkers_table.R
\name{make_biomarkers_table}
\alias{make_biomarkers_table}
\title{Make biomarkers table}
\usage{
make_biomarkers_table(
  data,
  biomarker_varnames = dplyr::pull(compile_biomarker_groups_table(), "biomarker"),
  biomarker_events_table,
  stratifying_var_names = "Gender"
)
}
\arguments{
\item{data}{a \code{\link[=data.frame]{data.frame()}} containing the columns specified by
\code{biomarker_varnames} as well as \code{stratifying_var_names}}

\item{biomarker_varnames}{a \link{character} vector matching a subset of
the column names in \code{data}}

\item{biomarker_events_table}{\link{data.frame} from
\code{\link[=construct_biomarker_events_table]{construct_biomarker_events_table()}}}

\item{stratifying_var_names}{a \link{character} vector specifying
variables to stratify by}
}
\value{
a \code{\link[flextable:flextable]{flextable::flextable()}}
}
\description{
Make biomarkers table
}
\examples{
biomarker_group_list <- list(
  "group 1" = c("Biomarker 1", "Biomarker 2"),
  "group 2" = c("Biomarker 3", "Biomarker 4"),
  "group 3" = "Biomarker 5"
)
biomarker_groups <- compile_biomarker_groups_table(
  dataset = sim_data,
  biomarker_group_list = biomarker_group_list
)

biomarker_varnames <-
  biomarker_groups |>
  dplyr::pull("biomarker")

biomarker_levels <-
  sim_data |>
  get_levels(biomarker_varnames)

biomarker_events_table <-
  construct_biomarker_events_table(
    biomarker_levels,
    biomarker_groups = biomarker_groups
  )

sim_data |> make_biomarkers_table(
  biomarker_events_table = biomarker_events_table,
  biomarker_varnames = biomarker_varnames,
  stratifying_var_names = "Sex"
)

}
