% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demographics_table.R
\name{make_demographics_table}
\alias{make_demographics_table}
\title{Make demographics table}
\usage{
make_demographics_table(data, strata_vars = "Gender")
}
\arguments{
\item{data}{todo}

\item{strata_vars}{todo}
}
\value{
a \link[flextable:flextable]{flextable::flextable}
}
\description{
Make demographics table
}
\examples{
test_data_v1 |> make_demographics_table()
}
