% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demographics_table.R
\name{make_demographics_table}
\alias{make_demographics_table}
\title{Make demographics table}
\usage{
make_demographics_table(data)
}
\arguments{
\item{data}{}
}
\description{
Make demographics table
}
