% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_reasons.R
\name{missingness_reasons}
\alias{missingness_reasons}
\title{Title}
\usage{
missingness_reasons(x, x.clean = clean_numeric(x, ...), ...)
}
\arguments{
\item{x}{the original vector to be turned into a numeric}

\item{x.clean}{the cleaned vector}

\item{...}{
  Arguments passed on to \code{\link[=clean_numeric]{clean_numeric}}
  \describe{
    \item{\code{NA_codes}}{numeric values that should be turned into to NAs}
    \item{\code{extra_codes}}{codes besides 777, 888, 999 to turn into NAs}
  }}
}
\description{
Title
}
