% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_dists.R
\name{model_dists}
\alias{model_dists}
\title{Title}
\usage{
model_dists(
  df,
  biomarker_varnames = attr(df, "biomarker_varnames"),
  biomarker_levels = lapply(df[, biomarker_varnames], FUN = levels),
  nlevs = sapply(biomarker_levels, length)
)
}
\arguments{
\item{df}{todo}

\item{biomarker_varnames}{todo}

\item{biomarker_levels}{todo}

\item{nlevs}{todo}
}
\value{
a \code{\link[=matrix]{matrix()}} indicating which levels each biomarker has
}
\description{
possibly not needed; keeping because it has an example of using attributes that I want to propagate across the codebase
}
