% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_permutation_pvalue.R
\name{permutation_test}
\alias{permutation_test}
\title{Compute permutation test p-value}
\usage{
permutation_test(permuted_test_stats, observed_test_stat)
}
\arguments{
\item{permuted_test_stats}{a \code{\link[=numeric]{numeric()}} vector}

\item{observed_test_stat}{a \code{\link[=numeric]{numeric()}} with length 1}
}
\value{
a \code{\link[=numeric]{numeric()}}
}
\description{
Compute permutation test p-value
}
