% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PF.R
\name{plot.PF}
\alias{plot.PF}
\title{Create positional variance diagram (PVD)}
\usage{
\method{plot}{PF}(PFs, size.y = 10, color_label = "Pr(stage)")
}
\arguments{
\item{PFs}{a "PF" object (created by \code{compute_position_frequencies()})}

\item{size.y}{size of biomarker event labels}

\item{color_label}{label for legend color scale}
}
\value{
a "PVD" object (extends \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}})
}
\description{
Create positional variance diagram (PVD)
}
