% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compact_pvd.R
\name{plot_compact_pvd}
\alias{plot_compact_pvd}
\title{Plot compact PVD
tmp function using list of extract_figs_from_pickle}
\usage{
plot_compact_pvd(
  figs,
  tile_height = 1,
  y_text_size = 9,
  facet_label_prefix = names(figs),
  legend.position = "none",
  scale_colors = c("red", "blue", "purple4", "darkgreen", "magenta"),
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=compact_pvd_data_prep]{compact_pvd_data_prep}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Plot compact PVD
tmp function using list of extract_figs_from_pickle
}
