% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compact_pvd.R
\name{plot_compact_pvd}
\alias{plot_compact_pvd}
\title{Plot compact PVD
tmp function using list of extract_figs_from_pickle}
\usage{
plot_compact_pvd(
  figs,
  tile_height = 1,
  y_text_size = 9,
  facet_label_prefix = names(figs),
  legend.position = "none",
  scale_colors = c("red", "blue", "purple4", "darkgreen", "magenta"),
  ...
)
}
\arguments{
\item{figs}{todo}

\item{tile_height}{todo}

\item{y_text_size}{todo}

\item{facet_label_prefix}{todo}

\item{legend.position}{todo}

\item{scale_colors}{todo}

\item{...}{
  Arguments passed on to \code{\link[=compact_pvd_data_prep]{compact_pvd_data_prep}}
  \describe{
    \item{\code{biomarker_order}}{a \link{character} vector specifying the
order of biomarkers to use}
  }}
}
\value{
a \link[ggplot2:ggplot]{ggplot2::ggplot}
}
\description{
Plot compact PVD
tmp function using list of extract_figs_from_pickle
}
