% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compact_pvd_est.R
\name{plot_compact_pvd_est}
\alias{plot_compact_pvd_est}
\title{@title Plot compact PVD with point estimates only
@export}
\usage{
plot_compact_pvd_est(
  figs,
  tile_height = 1,
  y_text_size = 9,
  facet_label_prefix = names(figs),
  legend.position = "none",
  scale_colors = c("red", "blue", "purple4", "darkgreen", "magenta")
)
}
\description{
@title Plot compact PVD with point estimates only
@export
}
