% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compact_pvd_est2.R
\name{plot_compact_pvd_est2}
\alias{plot_compact_pvd_est2}
\title{Plot SuStaIn Sequential Estimate}
\usage{
plot_compact_pvd_est2(
  figs,
  tile_height = 1,
  tile_width = 1,
  y_text_size = 9,
  legend.position = "none",
  scale_colors = c("red", "blue", "purple4", "darkgreen", "magenta")
)
}
\description{
Plot SuStaIn Sequential Estimate
}
