% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positional_variance.R
\name{plot_positional_var}
\alias{plot_positional_var}
\title{Plot positional variance diagram}
\usage{
plot_positional_var(
  results,
  samples_sequence = format_samples_sequence(results = results, biomarker_event_names =
    biomarker_event_names),
  samples_f = results$samples_f,
  n_samples = nrow(results$ml_subtype),
  score_vals,
  biomarker_labels = names(biomarker_levels),
  biomarker_groups = NULL,
  biomarker_levels = NULL,
  biomarker_events_table = get_biomarker_events_table(biomarker_levels),
  biomarker_event_names = pull(biomarker_events_table, biomarker_level),
  biomarker_plot_order = NULL,
  ml_f_EM = NULL,
  cval = FALSE,
  subtype_order = NULL,
  biomarker_order = NULL,
  title_font_size = 12,
  stage_font_size = 10,
  stage_label = "SuStaIn Stage",
  stage_rot = 0,
  stage_interval = 1,
  label_font_size = 10,
  label_rot = 0,
  cmap = "original",
  biomarker_colours = NULL,
  figsize = NULL,
  subtype_titles = NULL,
  separate_subtypes = FALSE,
  save_path = NULL,
  save_kwargs = NULL,
  synchronize_y_axes = FALSE,
  ...
)
}
\arguments{
\item{results}{pickle file contents}

\item{biomarker_labels}{biomarker labels}

\item{biomarker_groups}{biomarker groupings}

\item{biomarker_levels}{biomarker levels}

\item{biomarker_events_table}{table of biomarker events (excluding base level)}

\item{biomarker_event_names}{vector of biomarker event names}

\item{synchronize_y_axes}{}
}
\description{
Plot positional variance diagram
}
