% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positional_variance.R
\name{plot_positional_var}
\alias{plot_positional_var}
\title{Plot positional variance diagram}
\usage{
plot_positional_var(
  results,
  samples_sequence = results$samples_sequence,
  samples_f = results$samples_f,
  n_samples = nrow(results$ml_subtype),
  score_vals = build_score_vals(biomarker_levels),
  biomarker_labels = names(biomarker_levels),
  biomarker_groups = NULL,
  biomarker_levels = NULL,
  biomarker_events_table = get_biomarker_events_table(biomarker_levels),
  biomarker_event_names = pull(biomarker_events_table, biomarker_level),
  biomarker_plot_order = NULL,
  ml_f_EM = NULL,
  cval = FALSE,
  subtype_order = NULL,
  biomarker_order = NULL,
  title_font_size = 12,
  stage_font_size = 10,
  stage_label = "Sequential order",
  stage_rot = 0,
  stage_interval = 1,
  label_font_size = 10,
  label_rot = 0,
  cmap = "original",
  biomarker_colours = NULL,
  subtype_titles = NULL,
  separate_subtypes = FALSE,
  save_path = NULL,
  save_kwargs = NULL,
  synchronize_y_axes = FALSE,
  ...
)
}
\arguments{
\item{results}{todo}

\item{samples_sequence}{todo}

\item{samples_f}{todo}

\item{n_samples}{todo}

\item{score_vals}{todo}

\item{biomarker_labels}{todo}

\item{biomarker_groups}{biomarker groupings}

\item{biomarker_levels}{todo}

\item{biomarker_events_table}{todo}

\item{biomarker_event_names}{todo}

\item{biomarker_plot_order}{todo}

\item{ml_f_EM}{todo}

\item{cval}{todo}

\item{subtype_order}{todo}

\item{biomarker_order}{todo}

\item{title_font_size}{todo}

\item{stage_font_size}{todo}

\item{stage_label}{todo}

\item{stage_rot}{todo}

\item{stage_interval}{todo}

\item{label_font_size}{todo}

\item{label_rot}{todo}

\item{cmap}{a \link{character}}

\item{biomarker_colours}{a \link{character} vector of colors}

\item{subtype_titles}{todo}

\item{separate_subtypes}{todo}

\item{save_path}{todo}

\item{save_kwargs}{todo}

\item{synchronize_y_axes}{todo}

\item{...}{
  Arguments passed on to \code{\link[=plot.PF]{plot.PF}}
  \describe{
    \item{\code{PFs}}{a "PF" object (created by \code{compute_position_frequencies()})}
    \item{\code{size.y}}{size of biomarker event labels}
    \item{\code{color_label}}{label for legend color scale}
  }}
}
\value{
a \verb{"PVD.list} (a \link{list} of \code{PVD} objects from \code{\link[=plot.PF]{plot.PF()}})
}
\description{
Plot positional variance diagram
}
