% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd_lineplot.R
\name{pvd_lineplot}
\alias{pvd_lineplot}
\title{Plot change in Stage ranking}
\usage{
pvd_lineplot(
  figs,
  min_alpha = 0.25,
  max_alpha = 1,
  stage_alpha = 1,
  facet_labels = names(figs),
  text_size = 3.4,
  y_lab = "Sequential order",
  y_title_size = 12,
  y_text_size = 10,
  x_text_size = 12
)
}
\arguments{
\item{x_text_size}{}
}
\description{
Plot change in Stage ranking
}
