% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd_lineplot.R
\name{pvd_lineplot}
\alias{pvd_lineplot}
\title{Plot change in Stage ranking}
\usage{
pvd_lineplot(
  figs,
  min_alpha = 0.25,
  max_alpha = 1,
  stage_alpha = 1,
  facet_labels = names(figs),
  text_size = 3.4,
  y_lab = "Sequential order",
  y_title_size = 9,
  y_text_size = 8,
  x_text_size = 8
)
}
\arguments{
\item{figs}{a \link{list} of todo}

\item{facet_labels}{todo}

\item{text_size}{todo}

\item{y_title_size}{todo}

\item{y_text_size}{todo}

\item{x_text_size}{todo}

\item{alpha_nochange}{todo}
}
\description{
Plot change in Stage ranking
}
