% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd_lineplot.R
\name{pvd_lineplot}
\alias{pvd_lineplot}
\title{Plot change in Stage ranking}
\usage{
pvd_lineplot(
  figs,
  events_to_highlight = NULL,
  highlight_color = NA,
  min_alpha = 0.25,
  max_alpha = 1,
  stage_alpha = 1,
  facet_labels = names(figs),
  text_size = 3.4,
  y_lab = "Sequential order",
  y_title_size = 9,
  y_text_size = 8,
  x_text_size = 8,
  expand = 0.25
)
}
\arguments{
\item{figs}{a \link{list} of todo}

\item{events_to_highlight}{todo}

\item{highlight_color}{todo}

\item{min_alpha}{todo}

\item{max_alpha}{todo}

\item{stage_alpha}{todo}

\item{facet_labels}{todo}

\item{text_size}{todo}

\item{y_lab}{todo}

\item{y_title_size}{todo}

\item{y_text_size}{todo}

\item{x_text_size}{todo}

\item{expand}{\link{numeric} how much to pad the sides}
}
\description{
Plot change in Stage ranking
}
\examples{
output_path <-
  fs::path_package("extdata/sim_data", package = "fxtas")

figs <- extract_figs_from_pickle(
  output_folder = output_path,
  n = 3
)

figs[1:2] |> pvd_lineplot()
}
