% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvd_subtype_lineplot.R
\name{pvd_subtype_lineplot}
\alias{pvd_subtype_lineplot}
\title{Display Sequential Order of Events by Subtype}
\usage{
pvd_subtype_lineplot(
  figs,
  events_to_highlight,
  min_alpha = 0.25,
  max_alpha = 1,
  stage_alpha = 1,
  facet_labels = names(figs),
  text_size = 3.4,
  y_lab = "Sequential order",
  y_title_size = 9,
  y_text_size = 8,
  x_text_size = 8,
  subtype_x = c(1, 3, 5, 7),
  mult = 0.2,
  align_stage = TRUE,
  ...
)
}
\arguments{
\item{figs}{A \link{list} of todo}

\item{events_to_highlight}{todo}

\item{min_alpha}{todo}

\item{max_alpha}{todo}

\item{stage_alpha}{todo}

\item{facet_labels}{todo}

\item{text_size}{todo}

\item{y_lab}{todo}

\item{y_title_size}{todo}

\item{y_text_size}{todo}

\item{x_text_size}{todo}

\item{subtype_x}{Vector of x-axis value for the subtypes.}

\item{mult}{\link{numeric} vector}

\item{align_stage}{\link{logical} whether to align by FXTAS stage}

\item{...}{additional arguments passed to \link[ggbump:geom_bump]{ggbump::geom_bump}
Default = c(1, 1.15, 1.75, 2.35)}
}
\description{
Display Sequential Order of Events by Subtype
}
