% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_sex_differences.R
\name{report_sex_differences}
\alias{report_sex_differences}
\title{Report sex differences}
\usage{
report_sex_differences(table)
}
\arguments{
\item{table}{output from \code{\link[=make_biomarkers_table]{make_biomarkers_table()}}}
}
\value{
a \code{\link[=character]{character()}} string
}
\description{
Report sex differences
}
\examples{
biomarker_groups = compile_biomarker_groups_table()

biomarker_varnames =
  biomarker_groups |>
  pull("biomarker")

biomarker_levels =
trax_gp34_v1 |>
 dplyr::select(all_of(biomarker_varnames)) |>
 lapply(F = levels)

biomarker_events_table =
  construct_biomarker_events_table(
    biomarker_levels,
    biomarker_groups = biomarker_groups)

table = trax_gp34_v1 |> make_biomarkers_table(
  biomarker_events_table = biomarker_events_table,
  biomarker_varnames = biomarker_varnames
  )

  table |> report_sex_differences()

}
