% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_OSA.R
\name{run_OSA}
\alias{run_OSA}
\title{Title}
\usage{
run_OSA(
  prob_score,
  score_vals,
  SuStaInLabels,
  N_startpoints,
  N_S_max,
  N_iterations_MCMC,
  output_folder,
  dataset_name,
  use_parallel_startpoints = FALSE,
  seed = 1,
  plot = FALSE,
  N_CV_folds = 0,
  CV_fold_nums = 1:N_CV_folds,
  patient_data,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{\code{\link[=logical]{logical()}} indicating whether to print debugging information}
}
\description{
Title
}
