% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_run_info.R
\name{save_run_info}
\alias{save_run_info}
\title{Save run info to file}
\usage{
save_run_info(output_folder)
}
\arguments{
\item{output_folder}{a \link{character} indicating where to save info}
}
\value{
\link{TRUE}, invisibly
}
\description{
Save run info to file
}
\examples{
\dontrun{
save_run_info("output123/")
}
}
