% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_flextable_settings.R
\name{shared_flextable_settings}
\alias{shared_flextable_settings}
\title{Apply flextable settings to match Brain journal requirements}
\usage{
shared_flextable_settings(x)
}
\arguments{
\item{x}{a \link[flextable:flextable]{flextable::flextable}}
}
\value{
a modified \link[flextable:flextable]{flextable::flextable}
}
\description{
Apply flextable settings to match Brain journal requirements
}
