% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_formula.R
\name{stratified_formula}
\alias{stratified_formula}
\title{Title}
\usage{
stratified_formula(vars, stratum)
}
\arguments{
\item{stratum}{}
}
\description{
Title
}
\examples{
stratified_formula(c("Sepal.Length" ,"Sepal.Width"), "Species")
}
