% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visit1}
\alias{visit1}
\title{Visit 1}
\format{
\subsection{\code{visit1}}{

A data frame with nrow rows and ncol columns:
\describe{
\item{col1}{col1name}
}
}
}
\source{
\if{html}{\out{
<url>
}}
}
\usage{
visit1
}
\description{
First visit for each participant from \code{gp34}
}
\keyword{datasets}
